build fix

+++ tests/test_assert.h
@@ -1,12 +1,13 @@
 #pragma once
 
 #include <cmath>
+#include <cstdlib> // for ::exit
 
 #include <string>
 
 static void handle_assert(bool passed, const char* assert, const char* file, long line) {
     if (!passed) {
-        printf("Assertion failed %s in %s at line %d\n", assert, file, line);
+        printf("Assertion failed %s in %s at line %d\n", assert, file, (int)line);
         exit(-1);
     }
 }
